checkbugs: reduce false positives when checking for referenced bugs' milestones
Currently on develop we have patches referencing a number of bugs which
are in the 22.4 milestone. They were merged in from release-22, but are
not released yet (so not in master). So this check was picking them up,
even though there is nothing wrong with the patches being on develop.

It was also complaining about bug 1240809, which was originally patched
on develop but then cherry-picked and released in 22.1.

This restricts the check to only complain when the referenced bug's
milestone is not set, future_maint, or some newer version.