checkbugs: tweak and invert bug state checking logic
Mainly I wanted to adjust the logic so that it would accept a bug on
ASSIGNED when it has some patches MERGED and some NEW. Right now the
script was expecting the bug to be POST, not ASSIGNED, but that's overly
strict.

But in trying to make that change I thought it would be simpler to flip
the logic on its head: enumerate all the acceptable bug states,
depending on the state of the Gerrit patches.