checkbugs: get branch name even for detached HEAD
Jenkins, for whatever reason, doesn't check out the real git branch, it
checks out a detached HEAD pointing at the commit it wants to build.
That means we can't look at the local branch name to help guess whether
we are looking at a maintenance release or not.

Instead we can use this `git name-rev` trickery to figure out which
branch from the origin remote we are on. It will print out
"origin/release-22" or "origin/develop", assuming that we are on
a commit which matches some branch from origin.

It does mean that the script might guess wrongly if someone runs it from
their local git clone and their local branch is not up to date with
origin. But I don't think that matters.