dogfood-known-issues: show count of all jobs
This way we can get an idea of what proportion of them hit known issues.
Otherwise, spikes and troughs in the "all known issues" graph might give
the false impression that things are improving or getting worse when
they are just caused by variation in the number of jobs we ran in that
particular week.