checkbugs: don't complain about bugs which have been reverted and closed
Bug 1506430 had some patches merged, but then later reverted and the bug
was closed NOTABUG and milestone cleared. However checkbugs was still
complaining because it saw commits referencing that bug. It isn't smart
enough to understand that one of the commits was a reversion.

Let's use a special keyword in the devel whiteboard field of the bug,
just because that's simpler than trying to make checkbugs understand
what a revert commit is.